requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
blobfile=2.0.0=pypi_0
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.9.24=ha878542_0
cairo=1.16.0=hf32fb01_1
certifi=2022.9.24=pyhd8ed1ab_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.6=pypi_0
dbus=1.13.18=hb2f20db_0
einops=0.4.0=pypi_0
expat=2.4.1=h2531618_2
ffmpeg=4.0=hcdf2ecd_0
filelock=3.8.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
imageio=2.14.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=2.0.14=hd8c5072_2
jpeg=9d=h7f8727e_0
keras=2.6.0=pyhd3eb1b0_0
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h5101ec6_17
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.0=py37h06a4308_0
matplotlib-base=3.5.0=py37h3ed280b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mpi=1.0=openmpi
mpi4py=3.0.3=py37hd955b32_1
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
networkx=2.6.3=pypi_0
ninja=1.10.2=py37hd09550d_3
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openmpi=4.0.4=hdf1f1ad_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pcre=8.45=h295c915_0
pillow=8.4.0=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
pixman=0.40.0=h7f8727e_1
protobuf=3.19.1=py37h295c915_0
py-opencv=3.4.2=py37hb342d67_1
pycryptodomex=3.15.0=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.0=py3.7_cuda11.0.221_cudnn8.0.3_0
pytz=2022.1=pypi_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=py37h7f8727e_1
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
scikit-image=0.19.1=pypi_0
scipy=1.7.3=pypi_0
setuptools=58.0.4=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
tensorboardx=2.2=pyhd3eb1b0_0
thop=0.0.31-2005241907=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.5.4=pypi_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.7.0=py37
torchstat=0.0.7=pypi_0
torchvision=0.8.0=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.64.1=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.12=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
